+++ udata_front/theme/gouvfr/assets/js/admin.ts
@@ -1,21 +1,20 @@
 import { createApp } from "vue";
 import { createRouter, createWebHashHistory, type RouteRecordRaw } from 'vue-router';
-import { setupComponents } from "@etalab/data.gouv.fr-components";
+import { setupComponents } from "@datagouv/components";
 
 import "./dsfr.ts";
 // @ts-ignore
 import "@gouvfr/dsfr/dist/dsfr/dsfr.module";
 
-import Admin from "./pages/Admin/Admin.vue";
-
 import i18n from "./i18n.ts";
-import { admin_root, explorable_resources, schema_documentation_url, schema_validata_url, title } from "./config.ts";
+import { admin_root, schema_documentation_url, schema_validata_url, tabular_api_url, tabular_page_size, title } from "./config.ts";
 import Api from "./plugins/api.ts";
 import Auth from "./plugins/auth.ts";
 import InitSentry from "./sentry.ts";
 import Datasets from "./pages/Admin/Datasets/Datasets.vue";
 import Reuses from "./pages/Admin/Reuses/Reuses.vue";
 import Me from "./pages/Admin/Me/Me.vue";
+import Admin from "./pages/Admin/Admin.vue";
 import Members from "./pages/Admin/Members/Members.vue";
 import Profile from "./pages/Admin/Profile/Profile.vue";
 import Discussions from "./pages/Admin/Discussions/Discussions.vue";
@@ -23,11 +22,12 @@ import Discussions from "./pages/Admin/Discussions/Discussions.vue";
 setupComponents({
   admin_root,
   default_lang: i18n.global.locale.value,
-  explorable_resources,
   only_locales: i18n.global.locale.value,
   schema_documentation_url,
   schema_validata_url,
   show_copy_resource_permalink: true,
+  tabular_api_url,
+  tabular_page_size,
   title,
 });