+++ udata_front/theme/gouvfr/assets/js/helpers.ts
@@ -1,8 +1,8 @@
 // @ts-ignore
+import type { ResourceType } from "@datagouv/components";
 import RemoveMarkdown from "remove-markdown";
 import { readonly } from "vue";
 import { useI18n } from "vue-i18n";
-import type { ResourceType } from "./types";
 
 export const truncate = (val: string, length = 300) => {
   if (typeof val !== "string") return;
@@ -19,8 +19,6 @@ export function throwOnNever(_: never, message: string): never {
 }
 
 export const CLOSED_FORMATS = readonly(['pdf', 'doc', 'docx', 'word', 'xls', 'excel', 'xlsx'] as const);
-
-export const RESOURCE_TYPE = readonly(["main", "documentation", "update", "api", "code", "other"] as const);
 
 export const getResourceLabel = (type: ResourceType) => {
   const { t } = useI18n();